refactor: [EL-4914] place interactive tour provider and context structure to right place
