+++ src/[fsd]/app/providers/index.js
@@ -3,3 +3,5 @@ export * from './NodeCardProvider';
 export * from './AgentsStudioContext';
 export * from './ToolkitSocketContext.jsx';
 export * from './InstructionsInputRefContext.jsx';
+export * from './InteractiveTourProvider';
+export * from './InteractiveTourContext';
