+++ src/[fsd]/features/interactive-tours/index.js
@@ -1,4 +1,3 @@
-export { default as InteractiveTourProvider } from './ui/InteractiveTourProvider';
 export { useInteractiveTour } from './model/useInteractiveTour';
 export { useProposeTour } from './model/useProposeTour';
 export { useTourFromUrl } from './model/useTourFromUrl';
